ai_player/pyproject.toml
[project]
name = "poker-ai-player"
version = "0.1.0"
description = "PyTorch AI poker player service with gRPC interface"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "grpcio-health-checking>=1.60.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "structlog>=24.0.0",
    "protobuf>=5.0.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.4.0",
]
training = [
    "tensorboard>=2.15.0",
    "wandb>=0.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ai_player"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
